feat: support for input with mix of mono and multitrack audiostreams
Previously, encore would fail the job if the input had a mix of mono and
multitrack audiostreams, except if the first stream was multitrack. This feat
adds some support for inputs with a mix of mono and multitrackstreams,
where the first stream is a monostream. Given such an input, encore now uses
all monostreams that comes before the first multitrack stream, and ignore
all following audiostreams.

Signed-off-by: Gustav Grusell <gustav.grusell@eyevinn.se>