feat: implement segmented audio encoding
Introduces support for encoding audio separately from video in segments,
enabling better parallelization and throughput for large files.

There are now three modes for audio encoding when using segmented encoding:

1. ENCODE_WITH_VIDEO (default) -Audio is encoded together with each video segment
2. ENCODE_SEPARATELY_FULL - Audio is encoded once as a complete file,
then muxed with video segments
3. ENCODE_SEPARATELY_SEGMENTED - Audio is split into segments and encoded in parallel with video

Audio encoding mode can be selected with the 'encoreJob.audioEncodingMode' property.
A default value, used when the property is not specified on the job, can be configured
with the 'encore-settings.encoding.segmented-encoding.audio-encoding-mode' property.

For segmented audio encode, audio segment length can be configured with the
property 'encoreJob.audioSegmentLength' . Note that the segment length needs
to be a multiple of the audio frame duration for proper encoding. If audio
segment length is not specified, a suitable segment length close to 256s will
be calculated.

Segmented audio encode is only supported for aac and libfdk_aac codecs.

Signed-off-by: Gustav Grusell <gustav.grusell@eyevinn.se>