fix: handle errors during job initialization
Errors during job initialization were previously not handled, which
could result in a job being stuck in QUEUED state forever. Specifically,
a job with an invalid/unknown profile would be stuck in this state.
This fix now correcly handles error in these cases and sets the job
status to FAILED.

Signed-off-by: Gustav Grusell <gustav.grusell@eyevinn.se>